Rename sample module to sample_v1
